+++ fuzz/fuzz_targets/bitcoin/deserialize_transaction.rs
@@ -2,18 +2,18 @@ use honggfuzz::fuzz;
 
 fn do_test(data: &[u8]) {
     let tx_result: Result<bitcoin::blockdata::transaction::Transaction, _> =
-        bitcoin::consensus::encode::deserialize(data);
+        bitcoin::consensus::deserialize(data);
     match tx_result {
         Err(_) => {}
         Ok(mut tx) => {
-            let ser = bitcoin::consensus::encode::serialize(&tx);
+            let ser = bitcoin::consensus::serialize(&tx);
             assert_eq!(&ser[..], data);
             let len = ser.len();
             let calculated_weight = tx.weight().to_wu() as usize;
             for input in &mut tx.input {
                 input.witness = bitcoin::blockdata::witness::Witness::default();
             }
-            let no_witness_len = bitcoin::consensus::encode::serialize(&tx).len();
+            let no_witness_len = bitcoin::consensus::serialize(&tx).len();
             // For 0-input transactions, `no_witness_len` will be incorrect because
             // we serialize as segwit even after "stripping the witnesses". We need
             // to drop two bytes (i.e. eight weight). Similarly, calculated_weight is